requirements.txt
attrs==24.2.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.1.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
PyYAML==6.0.1
# Editable Git install with no remote (referral_tracker==0.1)
-e /Users/shanemartin/Desktop/PaperboyProjects/referral_tracker
requests==2.32.3
schedule==1.2.1
selenium==4.18.1
six==1.16.0
slack_sdk==3.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.27
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webdriver-manager==4.0.1
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
